encoder eccentricity calibration, current offset calibration

+++ software/firmware/src/focus_port.c
@@ -70,7 +70,7 @@ void HAL_ADCEx_InjectedConvCpltCallback(ADC_HandleTypeDef *hadc) {
     if(hadc == &hadc1) {
         HAL_GPIO_WritePin(GPIOB, GPIO_PIN_1, GPIO_PIN_SET);
 
-        const uint32_t enc = __HAL_TIM_GET_COUNTER(&htim2);
+        const uint16_t enc = __HAL_TIM_GET_COUNTER(&htim2);
         const uint32_t u = HAL_ADCEx_InjectedGetValue(&hadc1, ADC_INJECTED_RANK_1);
         const uint32_t v = HAL_ADCEx_InjectedGetValue(&hadc1, ADC_INJECTED_RANK_2);
         const uint32_t w = HAL_ADCEx_InjectedGetValue(&hadc1, ADC_INJECTED_RANK_3);
